Got client to work. also rearranged the folders to be more organized
