+++ index.js
@@ -1,48 +1,22 @@
-// This is the Web Server
+require('dotenv').config();
 const express = require('express');
-const server = express();
-
-// enable cross-origin resource sharing to proxy api requests
-// from localhost:3000 to localhost:4000 in local dev env
-const cors = require('cors');
-server.use(cors());
-
-// create logs for everything
+const app = express();
 const morgan = require('morgan');
-server.use(morgan('dev'));
+const router = require('./api');
+const cors = require('cors');
 
-// handle application/json requests
-server.use(express.json());
+app.use(cors());
+app.use(morgan("dev")); 
+app.use(express.json());
 
-// here's our static files
-const path = require('path');
-server.use(express.static(path.join(__dirname, 'build')));
+app.use('/api', router);
 
-// here's our API
-server.use('/api', require('./api'));
+const PORT = process.env.PORT || 5432;
 
-// by default serve up the react app if we don't recognize the route
-server.use((req, res, next) => {
-  res.sendFile(path.join(__dirname, 'build', 'index.html'));
-});
+const client = require('./db/client');
+client.connect();
 
-// bring in the DB connection
-const { client } = require('./db');
 
-// connect to the server
-const PORT = process.env.PORT || 4000;
-
-// define a server handle to close open tcp connection after unit tests have run
-const handle = server.listen(PORT, async () => {
-  console.log(`Server is running on ${PORT}!`);
-
-  try {
-    await client.connect();
-    console.log('Database is open for business!');
-  } catch (error) {
-    console.error('Database is closed for repairs!\n', error);
-  }
-});
-
-// export server and handle for routes/*.test.js
-module.exports = { server, handle };
+app.listen(PORT, () => {
+    console.log(`Server is up and running on port ${PORT}`)
+})
