+++ package.json
@@ -24,9 +24,9 @@
     "jest": "^24.9.0",
     "morgan": "^1.10.0",
     "pg": "^8.2.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.6.1",
     "react-scripts": "^5.0.1"
   },
   "devDependencies": {
