+++ public/index.html
@@ -2,11 +2,12 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Full-Stack App</title>
+    <link rel="stylesheet" href="styles.css" />
+    <title>Hello World testing</title>
   </head>
   <body>
     <div id="root"></div>
   </body>
 </html>
-
