+++ src/App.js
@@ -1,0 +1,20 @@
+import React, { useState, useEffect } from 'react';
+import { Route, BrowserRouter as Routes, useNavigate } from 'react-router-dom';
+
+
+
+
+
+const App = () => {
+
+    return (      
+            <div className="app-container">
+              <h1>Hello, World! This is a quick test to load homepage</h1>
+
+            </div>
+    
+    )
+}
+
+
+export default App;
